+++ idea/testData/refactoring/move/kotlin/moveFile/callableReferences/before/c/util.kt
@@ -1,0 +1,3 @@
+package c
+
+fun f4() = 4